+++ RELEASES.md
@@ -9,6 +9,7 @@ Version 0.9.6 (2024-02-22)
 * AGNOS 9
 * comma body streaming and controls over WebRTC
 * Improved fuzzy fingerprinting for many makes and models
+* Alpha longitudinal support for new Toyota models
 * Chevrolet Equinox 2019-22 support thanks to JasonJShuler and nworb-cire!
 * Dodge Duranago 2020-21 support
 * Hyundai Staria 2023 support thanks to sunnyhaibin!